+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.31.0]
 ### Added
 - NumberAsString option
+- Queue.PurgeExpired
 ### Changed
 - By default, return Number for numbers (not string)
+- Always run session Init functions
 
 ## [v0.30.2]
 ### Changed